Hook v8 instead of pthreads
This iteration hooks v8's thread abstraction instead of pthreads
directly. It has the advantage of being much less platform-specific and
probably a lot more stable. It has the disadvantage that it will only
work if you have v8 compiled as a shared library, which is a deal
breaker. If only there were some way I could get this code into node
without forcing a recompile..