Fibers based on pthreads instead of ucontext
This code is a little scrappy.. this is just an experiment I wanted to
run to quantify the advantages of using makecontext over pthreads. It
was possible to implement the current fibers library using only pthreads
with cond signaling (similar to v8cgi's implementation), however the
performance wasn't close enough to ucontext's performance.

In one case I wasn't able to utilize 100% CPU usage and saw a
performance hit there. In another case I was hitting >115% CPU usage
(multi-core) however much of this was spent in the scheduler, and the
test in question ran about half as fast.

The advantage is that you no longer need the node-fibers script with
this implementation, which means it's more portable and there's no more
LD_LIBRARY_PRELOAD'ing black magic which is desirable. Perhaps this code
could find some use after a clean up and more experimentation.